+++ test/widget_tests/flower_menu_widget_test.dart
@@ -25,119 +25,6 @@ void main() {
       //middleware: createStoreMiddleware(),
     );
 
-    testWidgets("sets size correctly around the flower",
-        (WidgetTester tester) async {
-      await setUpWidget(tester, store);
-
-      await tester.pump();
-      await tester.pump();
-
-      // Find first container inside flower (ie. immediate child)
-      Container cont = find.descendant(of: find.byType(FlowerMenu), matching: find.byType(Container)).evaluate().first.widget;
-      FlowerMenu flower = find.byType(FlowerMenu).evaluate().first.widget;
-
-
-      // Check that it's padding property is set correctly
-      expect(
-          cont.constraints,
-          BoxConstraints.expand(
-              width: flower.flowerSize, height: flower.flowerSize));
-    });
-
-    testWidgets("sets insets correctly around the flower to center it",
-        (WidgetTester tester) async {
-      await setUpWidget(tester, store);
-
-      Container cont = find.descendant(of: find.byType(FlowerMenu), matching: find.byType(Container)).evaluate().elementAt(1).widget;
-      FlowerMenu flower = find.byType(FlowerMenu).evaluate().first.widget;
-
-      // width and height in a Container translate into BoxConstraints
-      expect(
-          cont.padding,
-          EdgeInsets.only(left: flower.flowerSize / 2, top: flower.flowerSize / 2));
-    });
-
-    testWidgets("has a Stack widget", (WidgetTester tester) async {
-      await setUpWidget(tester, store);
-
-      Container cont = find.descendant(of: find.byType(FlowerMenu), matching: find.byType(Container)).evaluate().elementAt(1).widget;
-
-      expect(cont.child.runtimeType, Stack);
-    });
-
-    testWidgets("contains 11 petals", (WidgetTester tester) async {
-      await setUpWidget(tester, store);
-
-      expect(find.byType(FlowerPetal).evaluate().length, 11);
-    });
-
-    testWidgets("assigns correct parameters to Flower Petals",
-        (WidgetTester tester) async {
-      await setUpWidget(tester, store);
-
-      var petals = find.byType(FlowerPetal).evaluate();
-      FlowerMenu flower = find.byType(FlowerMenu).evaluate().first.widget;
-
-      FlowerPetal petal;
-
-      // Might wanna separate into several tests. I left it like this to not
-      // have to generate the widget 11 times.
-      petal = petals.elementAt(0).widget;
-      expect(petal.progress, 50.0);
-      expect(petal.angle, 0.0);
-      expect(petal.color, Colors.red);
-
-      petal = petals.elementAt(1).widget;
-      expect(petal.progress, 50.0);
-      expect(petal.angle, (2 / 11) * pi);
-      expect(petal.color, Colors.blue);
-
-      petal = petals.elementAt(2).widget;
-      expect(petal.progress, 50.0);
-      expect(petal.angle, (4 / 11) * pi);
-      expect(petal.color, Colors.lightBlue);
-
-      petal = petals.elementAt(3).widget;
-      expect(petal.progress, 50.0);
-      expect(petal.angle, (6 / 11) * pi);
-      expect(petal.color, Colors.lightGreen);
-
-      petal = petals.elementAt(4).widget;
-      expect(petal.progress, 50.0);
-      expect(petal.angle, (8 / 11) * pi);
-      expect(petal.color, Colors.indigo);
-
-      petal = petals.elementAt(5).widget;
-      expect(petal.progress, 50.0);
-      expect(petal.angle, (10 / 11) * pi);
-      expect(petal.color, Colors.amber);
-
-      petal = petals.elementAt(6).widget;
-      expect(petal.progress, 50.0);
-      expect(petal.angle, (12 / 11) * pi);
-      expect(petal.color, Colors.blueGrey);
-
-      petal = petals.elementAt(7).widget;
-      expect(petal.progress, 50.0);
-      expect(petal.angle, (14 / 11) * pi);
-      expect(petal.color, Colors.deepOrange);
-
-      petal = petals.elementAt(8).widget;
-      expect(petal.progress, 50.0);
-      expect(petal.angle, (16 / 11) * pi);
-      expect(petal.color, Colors.lime);
-
-      petal = petals.elementAt(9).widget;
-      expect(petal.progress, 50.0);
-      expect(petal.angle, (18 / 11) * pi);
-      expect(petal.color, Colors.purple);
-
-      petal = petals.elementAt(10).widget;
-      expect(petal.progress, 50.0);
-      expect(petal.angle, (20 / 11) * pi);
-      expect(petal.color, Colors.green);
-    });
-
     testWidgets('has an invisible button with correct (small) size initially', (WidgetTester tester) async  {
       await setUpWidget(tester, store);
